requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.5
django-cloudinary-storage==0.3.0
djangorestframework
django-cors-headers
dj-database-url==2.3.0
djangorestframework-simplejwt
filelock==3.17.0
Flask==3.1.0
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==25.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
virtualenv==20.29.2
Werkzeug==3.1.3
whitenoise==6.9.0
Pillow
django-cloudinary-storage
cloudinary
channels
channels_redis